Feat: Pickup loot and case insensitive inputs
Using [xX] in the case statements input can be lower or upper, preventing a needless error. Added the array moves to track the moves the player makes, a stack would be preferable for current use as only the last value is needed for the revert moves and repeat moves functions.